Clarify the project README and improve starter code
Add code to the project to ease students into the task of writing a
solution to the primary challenge of generating markdown.

Include clearer language in the README file to avoid confusion and links
to content that does not pertain directly to the project tasks.

Include the project tasks in the README so that students who use project
as starter code can reference the instructions instead of having to
visit another website outside of the project files.

Simplify the usage of the "marked" library by including it as a global
script in the <head> of the "index.html" file, instead of using an NPM
package. This removes the need to add the "node_modules" folder as a
directory which can be served by the web server.

Remove the package "node-static" in favor of the package "serve" as the
later is more modern and better maintained than the former.

Update the project dependencies to remove security vulnerabilities and
keep the project up to date with the packages as maintained by the
open source community.